ref: move decimal256ext to its own file in packages
