+++ packages/euclid/src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod fund_manager;
+pub mod math;
 pub mod pagination;
 pub mod tx;
